backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
langgraph
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-ollama
chromadb
psycopg2-binary
regex
sqlglot
python-multipart
email-validator
pymysql
cryptography
motor
dnspython
